requirements.txt
fastmcp[sse]
sqlalchemy
psycopg2-binary
python-dotenv
rapidfuzz